Migrate Layout Extras - All shared components migrated to _shared directory - Single use components migrated to file used within
